Update Code Block Formatting in Markdown
This commit updates the markdown format for code blocks in our documentation. Previously, we were using the 'bash' markdown, which did not provide optimal readability for our Go code examples. In this update, we've changed the markdown to 'go' for a better visual representation and improved readability of the code blocks.