Refactor: Removed Comment and Streamlined ValidateFileIntegrity Function (#122)
Removed the comment '// This should be in the same method or function as the switch statement.' as the design choice was deliberate. The intention behind passing the hash value to the ValidateFileIntegrity function was to maintain simplicity and avoid bloating the function. Introducing other values would unnecessarily increase its complexity.

Co-authored-by: Marco Villarruel <marco@wildberry.us>